BOT V1.1

+++ bot.py
@@ -1,4 +1,5 @@
 import logging
+from DBMS import log_activity
 import os
 import asyncio
 from shazamio import Shazam
@@ -37,6 +38,9 @@ async def helP(update:Update,context:ContextTypes.DEFAULT_TYPE):
     await update.message.reply_text(
     "use /music and type song name or URL\n\n"
     "use /video and type video name or URL\n\n"
+    "or if you dont know the name of your song\n"
+    "send me audio file have part of the song\n"
+    "I will recognazed it and upload it to you"
     )
 
 async def about(update:Update,contxt:ContextTypes.DEFAULT_TYPE):
@@ -52,6 +56,7 @@ async def music(update: Update, context: ContextTypes.DEFAULT_TYPE):
         return
     
     song_name = " ".join(context.args)
+    file_path = None
 
     if isURL(song_name) == True:
         status_message = await update.message.reply_text("🔗 Searching for the URL...")
@@ -62,6 +67,11 @@ async def music(update: Update, context: ContextTypes.DEFAULT_TYPE):
     try:
 
         title, artist, file_path = await asyncio.to_thread(download_audio,song_name)
+        try:
+            await log_activity(update.effective_user.id, update.effective_user.username, "download", song_name)
+
+        except Exception as e:
+            print(f"Logging failed: {e}")
 
         await status_message.edit_text("🎵 Sending the audio file, please wait...")
                 
@@ -73,7 +83,10 @@ async def music(update: Update, context: ContextTypes.DEFAULT_TYPE):
                 read_timeout=120,  
                 write_timeout=120
             )
-                
+        try:
+            await log_activity(update.effective_user.id, update.effective_user.username, "upload", song_name)
+        except Exception as e:
+            print(f"Logging failed: {e}")
     except TimedOut:
         
         pass 
@@ -97,6 +110,8 @@ async def video(update:Update,context:ContextTypes.DEFAULT_TYPE):
         return
     
     video_name = " ".join(context.args)
+    file_path = None
+    thumb_path = None
 
     if isURL(video_name) == True:
         status_message = await update.message.reply_text("🔗 Searching for the URL...")
@@ -105,8 +120,13 @@ async def video(update:Update,context:ContextTypes.DEFAULT_TYPE):
 
     try:
 
-        file_path, info = download_video(video_name)
-        thumb_path = generate_thumbnail(file_path)
+        file_path, info = await asyncio.to_thread(download_video, video_name)
+        thumb_path = await asyncio.to_thread(generate_thumbnail, file_path)
+
+        try:
+            await log_activity(update.effective_user.id, update.effective_user.username, "download", video_name)
+        except Exception as e:
+            print(f"Logging failed: {e}")
 
         await status_message.edit_text(" 🎥 Sending the video file, please wait...")
                 
@@ -121,7 +141,11 @@ async def video(update:Update,context:ContextTypes.DEFAULT_TYPE):
                 read_timeout=120,  
                 write_timeout=120,
             )
-                
+        try:
+            await log_activity(update.effective_user.id, update.effective_user.username, "upload", video_name)
+        except Exception as e:
+            print(f"Logging failed: {e}")
+
     except TimedOut:
         
         pass 
@@ -141,39 +165,65 @@ async def video(update:Update,context:ContextTypes.DEFAULT_TYPE):
             os.remove(thumb_path)
 
 async def song_recognazed(update:Update,context:ContextTypes.DEFAULT_TYPE):
-    ...
+
     status = await update.message.reply_text("🎧 Listening...")
 
-    audio = update.message.audio
-    file = await audio.get_file()
-    local_path = cdir("downloads")
-    local_path = os.path.join(local_path, file.file_unique_id) + ".mp3"
-    file = await audio.get_file()
-    await file.download_to_drive(local_path,read_timeout=120,write_timeout=120)
+    try:
 
-    shazam = Shazam()
-    result = await shazam.recognize(local_path)
+        audio = update.message.audio
+        file = await audio.get_file()
+        local_path = cdir("downloads")
+        local_path = os.path.join(local_path, file.file_unique_id) + ".mp3"
+        file = await audio.get_file()
+        await file.download_to_drive(local_path,read_timeout=120,write_timeout=120)
 
-    track = result.get("track")
-    if not track:
-        await status.edit_text("Couldn't recognize that one — try a clearer clip.")
-        os.remove(local_path)
-        return
+    except Exception as e:
+        await status.edit_text(f"something went wrong with download the audio file {e}")
+
+    try:
+        shazam = Shazam()
+        result = await shazam.recognize(local_path)
+
+        track = result.get("track")
+        if not track:
+            await status.edit_text("Couldn't recognize that one — try a clearer clip.")
+            os.remove(local_path)
+            return
+    except Exception as e:
+        await status.edit_text(f"something went wrong with recognazed the song {e}")
 
     title = track["title"]
     artist = track["subtitle"]
     await status.edit_text(f"Found: song {title}\n by {artist} \n seending  the full song...")
 
-    title, artist, file_path = download_audio(title)
- 
-    with open(file_path, "rb") as audio_file:
-        await update.message.reply_audio(audio=audio_file,
-            title=title,
-            performer=artist,
-            read_timeout=120,
-            write_timeout=120
-            )
+    title, artist, file_path = await asyncio.to_thread( download_audio, title)
+
+    try:
+        await log_activity(update.effective_user.id, update.effective_user.username, "download", title)
+    except Exception as e:
+        print(f"Logging failed: {e}")
+    
+    try:
+
+        with open(file_path, "rb") as audio_file:
+            await update.message.reply_audio(audio=audio_file,
+                title=title,
+                performer=artist,
+                read_timeout=120,
+                write_timeout=120
+                )
+    except Exception as e:
+        await status.edit_text(f"something went wrong with upload the song {e}")
+
+    try:
+        await log_activity(update.effective_user.id, update.effective_user.username, "upload", title)
+    except Exception as e:
+        print(f"Logging failed: {e}")
+
     await status.delete()
     
-    os.remove(local_path)
-    os.remove(file_path)
+    if local_path and os.path.exists(local_path):
+        os.remove(local_path)
+    if file_path and os.path.exists(file_path):
+        os.remove(file_path)
+
